# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TehTruth Dev Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List all tracked tokens
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Register a deployed token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenBody"
      responses:
        "201":
          description: Token saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{id}:
    get:
      operationId: getToken
      tags: [tokens]
      summary: Get token by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "404":
          description: Not found
    delete:
      operationId: deleteToken
      tags: [tokens]
      summary: Remove a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /tools/encode:
    post:
      operationId: encodeCalldata
      tags: [tools]
      summary: ABI-encode function calldata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncodeCalldataBody"
      responses:
        "200":
          description: Encoded calldata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncodedResult"
  /tools/decode:
    post:
      operationId: decodeCalldata
      tags: [tools]
      summary: Decode hex calldata with ABI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecodeCalldataBody"
      responses:
        "200":
          description: Decoded result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecodedResult"
  /tools/checksum:
    post:
      operationId: checksumAddress
      tags: [tools]
      summary: Compute EIP-55 checksum address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecksumBody"
      responses:
        "200":
          description: Checksummed address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecksumResult"
  /tools/keccak256:
    post:
      operationId: keccak256Hash
      tags: [tools]
      summary: Compute Keccak256 hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashBody"
      responses:
        "200":
          description: Hash result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResult"
...
  /session-keys:
    get:
      operationId: listSessionKeys
      tags: [session-keys]
      summary: List session keys
      responses:
        "200":
          description: List of session keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionKey"
    post:
      operationId: createSessionKey
      tags: [session-keys]
      summary: Create a session key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionKeyBody"
      responses:
        "201":
          description: Session key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionKey"
...
  /session-keys/{id}:
    get:
      operationId: getSessionKey
      tags: [session-keys]
      summary: Get session key by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session key details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionKey"
        "404":
          description: Not found
    patch:
      operationId: updateSessionKey
      tags: [session-keys]
      summary: Update or revoke a session key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionKeyBody"
      responses:
        "200":
          description: Session key updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionKey"
        "404":
          description: Not found
    delete:
      operationId: deleteSessionKey
      tags: [session-keys]
      summary: Remove session key record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    Token:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        name:
          type: string
        symbol:
          type: string
        decimals:
          type: integer
        tokenType:
          type: string
        chainId:
          type: integer
        totalSupply:
          type: ["string", "null"]
        deployTxHash:
          type: ["string", "null"]
        deployer:
          type: ["string", "null"]
        wasPrivate:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - address
        - name
        - symbol
        - decimals
        - tokenType
        - chainId
        - wasPrivate
        - createdAt
    CreateTokenBody:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        symbol:
          type: string
        decimals:
          type: integer
        tokenType:
          type: string
        chainId:
          type: integer
        totalSupply:
          type: ["string", "null"]
        deployTxHash:
          type: ["string", "null"]
        deployer:
          type: ["string", "null"]
        wasPrivate:
          type: boolean
      required:
        - address
        - name
        - symbol
        - tokenType
        - chainId
...
    EncodeCalldataBody:
      type: object
      properties:
        functionSignature:
          type: string
        params:
          type: array
          items:
            type: string
      required:
        - functionSignature
        - params
    EncodedResult:
      type: object
      properties:
        encoded:
          type: string
        selector:
          type: string
      required:
        - encoded
        - selector
    DecodeCalldataBody:
      type: object
      properties:
        calldata:
          type: string
        abi:
          type: array
          items:
            type: object
      required:
        - calldata
        - abi
    DecodedResult:
      type: object
      properties:
        functionName:
          type: string
        params:
          type: array
          items:
            type: object
      required:
        - functionName
        - params
    ChecksumBody:
      type: object
      properties:
        address:
          type: string
      required:
        - address
    ChecksumResult:
      type: object
      properties:
        checksummed:
          type: string
        isValid:
          type: boolean
      required:
        - checksummed
        - isValid
    HashBody:
      type: object
      properties:
        input:
          type: string
      required:
        - input
    HashResult:
      type: object
      properties:
        hash:
          type: string
      required:
        - hash
...
    SessionKey:
      type: object
      properties:
        id:
          type: integer
        walletAddress:
          type: string
        sessionPublicKey:
          type: string
        chainId:
          type: integer
        label:
          type: ["string", "null"]
        permissions:
          type: ["object", "null"]
        allowedTargets:
          type: ["array", "null"]
          items:
            type: string
        spendLimit:
          type: ["string", "null"]
        validAfter:
          type: ["string", "null"]
          format: date-time
        validUntil:
          type: ["string", "null"]
          format: date-time
        isRevoked:
          type: boolean
        txHash:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - walletAddress
        - sessionPublicKey
        - chainId
        - isRevoked
        - createdAt
    CreateSessionKeyBody:
      type: object
      properties:
        walletAddress:
          type: string
        sessionPublicKey:
          type: string
        chainId:
          type: integer
        label:
          type: ["string", "null"]
        permissions:
          type: ["object", "null"]
        allowedTargets:
          type: ["array", "null"]
          items:
            type: string
        spendLimit:
          type: ["string", "null"]
        validAfter:
          type: ["string", "null"]
          format: date-time
        validUntil:
          type: ["string", "null"]
          format: date-time
        txHash:
          type: ["string", "null"]
      required:
        - walletAddress
        - sessionPublicKey
        - chainId
    UpdateSessionKeyBody:
      type: object
      properties:
        isRevoked:
          type: boolean
        label:
          type: ["string", "null"]
...