# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TehTruth Dev Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  id: number;
  address: string;
  name: string;
  symbol: string;
  decimals: number;
  tokenType: string;
  chainId: number;
  /** @nullable */
  totalSupply?: string | null;
  /** @nullable */
  deployTxHash?: string | null;
  /** @nullable */
  deployer?: string | null;
  wasPrivate: boolean;
  createdAt: Date;
}